pyproject.toml
[project]
name = "langchainlearn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=1.1.10",
    "langchain[anthropic,google-genai,mistralai,openai]>=1.2.10",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
